Reimplement company number sanitizer
